Add `RouteMap` composable and reintegrate `os-map` module
Introduce `RouteMap` composable utilizing MapLibre Compose for dynamic GeoJSON rendering. Reinstate and configure `:os-map` module with required dependencies, including `maplibre-compose` and `platformtools`. Replace direct `MaplibreMap` calls in `PlanningTransportRoutePreviewPage` with `RouteMap`. Update build scripts for enabling Android targets and iOS framework support.
